backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-slugify
slowapi
